haystack_service/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
haystack-ai>=2.10.3
pgvector-haystack>=0.1.0
unstructured[pdf]>=0.11.0  # Add more file types with unstructured[all-docs] if needed
python-multipart>=0.0.6
httpx>=0.24.0
psycopg2-binary>=2.9.6
pydantic>=2.0.0
unstructured-fileconverter-haystack>=0.1.0